+++ src/components/About/Skill.astro
@@ -1,0 +1,6 @@
+<div>
+  <div>
+    <slot name="tooltip" />
+  </div>
+  <slot />
+</div>
