+++ src/components/Head/BaseHead.astro
@@ -1,8 +1,4 @@
-<script defer type="module" src={Astro.resolve("./tippy.ts")}></script>
 <style global lang="scss">
-  @use "tippy.js/dist/tippy.css";
-  @use "tippy.js/dist/svg-arrow.css";
-  @use "tippy.js/animations/shift-away.css";
   @import url('https://api.fontshare.com/css?f[]=chillax@200,300,400,600,700,1&display=swap');
 </style>
 <style global>
@@ -73,15 +69,5 @@
       transform: translateY(0);
     }
   }
-
-  .tippy-box[data-theme~='space'] {
-    background-color: rgba(50, 50, 50, 0.1);
-    color: white;
-  }
-
-  .tippy-box[data-theme~='emotion'] {
-    background: linear-gradient(130deg, #373b44, #4286f4);
-    box-shadow: 2px 2px 10px 5px rgba(37, 112, 173, 0.5);
-    font-weight: 400;
-  }
 </style>
+<script defer type="module" src={Astro.resolve('./blobity.ts')} />
