+++ src/components/Navbar/Navbar.astro
@@ -161,13 +161,13 @@ Burger king
 	</nav>
 </div>
 <div class="top">
-	<a href="/" class="logo" main-tooltip data-tippy-content="Rishikesh 'Blaine' Anand"><img src="/assets/logo.svg" width="100%" height="100%" alt="Logo"/></a>
+	<a href="/" class="logo" main-tooltip data-blobity-tooltip="Rishikesh 'Blaine' Anand"><img src="/assets/logo.svg" width="100%" height="100%" alt="Logo"/></a>
 	<div class="right">
-		<a type="button" id='toggleScene' class="toggle-off inherit" main-tooltip data-tippy-content="Hide HUD<br />(Allows interaction with content)">
+		<a type="button" id='toggleScene' class="toggle-off inherit" main-tooltip data-blobity-tooltip="Hide HUD<br />(Allows interaction with content)">
       <Icon name="fa-solid:toggle-off" classList="off" />
       <Icon name="fa-solid:toggle-on" classList="on" />
     </a>
-		<a class="inherit" href="https://github.com/Blakeinstein" main-tooltip data-tippy-content="Github Profile<pre>                       </pre>@Blakeinstein"><Icon name="ci:github" /></a>
+		<a class="inherit" href="https://github.com/Blakeinstein" main-tooltip data-blobity-tooltip="Github Profile<pre>                       </pre>@Blakeinstein"><Icon name="ci:github" /></a>
 		<div class="burger" />
 	</div>
 </div>
