+++ src/components/Navbar/navbar.ts
@@ -1,6 +1,3 @@
-import type { Options } from 'blobity/lib/Blobity';
-import Blobity from 'blobity';
-
 import anime from "animejs/lib/anime.es.js";
 import Swup from 'swup';
 import SwupSlideTheme from "@swup/slide-theme";
@@ -111,15 +108,3 @@ swup.preloadPages();
 
 init();
 swup.on('contentReplaced', init);
-
-const blobityOptions: Partial<Options> = {
-	licenseKey: 'blobity-license-key',
-	color: '#e3f2fd',
-	opacity: 0.3,
-	size: 28,
-	dotColor: '#eeeeee',
-	focusableElements: '.burger, a, a svg, #noSpamDiv>label, button, input, textarea, select, .swup-transition-link',
-	mode: 'slow',
-}
-
-const blobity = new Blobity(blobityOptions);
