+++ src/components/Social/Link.astro
@@ -35,7 +35,7 @@ li {
 </style>
 
 <li>
-  <a emotion-tool-tip data-tippy-content={tippyContent} href={href}>
+  <a data-blobity-tooltip={tippyContent} href={href}>
     <BrandIcon name={icon} />
   </a>
 </li>
