+++ src/layouts/BaseLayout.astro
@@ -5,7 +5,6 @@ import Navbar from '../components/Navbar/Navbar.astro';
 import BaseHead from '../components/Head/BaseHead.astro';
 import Content from '../components/Content.astro';
 
-
 export interface Props {
   title: string;
   description: string;
