+++ tsconfig.json
@@ -1,3 +1,6 @@
 {
-  "moduleResolution": "node"
+  "compilerOptions": {
+    "moduleResolution": "node"
+  },
+  "include": ["env.d.ts"]
 }
